requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.3
python-multipart>=0.0.9
aiofiles>=23.2.1

# Auth & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.0

# Hardware
pyserial>=3.5

# Database
sqlalchemy>=2.0.30
alembic>=1.13.0

# ML / Detection
ultralytics>=8.2.0
fast-plate-ocr[onnx]>=1.1.0
opencv-python>=4.9.0
torch>=2.3.0
torchvision>=0.18.0

# Utils
numpy>=1.26.0
pillow>=10.3.0
pandas>=2.2.0
pyyaml>=6.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
